VC2015 warning fixes

+++ source/Win32/i_opndir.cpp
@@ -211,7 +211,7 @@ struct dirent *readdir(DIR *dirp)
       /* Successfully got an entry. Everything about the file is
        * already appropriately filled in except the length of the
        * file name. */
-      dirp->dd_dir.d_namlen = _tcslen(dirp->dd_dta.name);
+      dirp->dd_dir.d_namlen = static_cast<unsigned short>(_tcslen(dirp->dd_dta.name));
       _tcscpy(dirp->dd_dir.d_name, dirp->dd_dta.name);
       return &dirp->dd_dir;
    }
